feat/chat-panel-multiselect-tests: introduced assertion classes
