+++ apps/chat-e2e/.eslintrc.json
@@ -14,7 +14,8 @@
       "rules": {
         "playwright/no-conditional-in-test": "off",
         "playwright/no-conditional-expect": "off",
-        "playwright/no-skipped-test": "off"
+        "playwright/no-skipped-test": "off",
+        "playwright/expect-expect": "off"
       }
     }
   ]
