+++ apps/chat-e2e/src/testData/expectedMessages.ts
@@ -297,9 +297,12 @@ export enum ExpectedMessages {
   copiedContentIsValid = 'Copied content is valid',
   folderIsChecked = 'Folder is checked',
   folderContentIsPartiallyChecked = 'Folder content is partially checked',
-  conversationIsChecked = 'Conversation is checked',
+  entityIsChecked = 'Entity is checked',
   folderIsNotChecked = 'Folder is not checked',
-  conversationIsNotChecked = 'Conversation is not checked',
+  entityIsNotChecked = 'Entity is not checked',
   folderEntityBackgroundColorIsValid = 'Folder entity background color is valid',
+  buttonIsVisible = 'Button is visible',
   buttonIsNotVisible = 'Button is not visible',
+  folderEntityIsNotVisible = 'Folder entity is not visible',
+  folderEntityIsVisible = 'Folder entity is visible',
 }
