+++ apps/chat-e2e/src/testData/types.ts
@@ -1,0 +1,6 @@
+export interface ChatBarEntity {
+  name: string;
+  index?: number;
+}
+
+export type ElementState = 'visible' | 'hidden';
