+++ apps/chat-e2e/src/tests/desktopAuth.ts
@@ -9,7 +9,6 @@ const usernames = process.env
   .slice(0, +config.workers! + 2);
 
 for (let i = 0; i < usernames.length; i++) {
-  // eslint-disable-next-line playwright/expect-expect
   test(`Authenticate user: ${usernames[i]}`, async ({
     page,
     loginPage,
