+++ apps/chat-e2e/src/ui/webElements/baseConversation.ts
@@ -104,4 +104,12 @@ export class BaseConversation extends SideBarEntities {
   public async getConversationBackgroundColor(name: string, index?: number) {
     return this.getEntityBackgroundColor(this.entitySelector, name, index);
   }
+
+  public getConversationCheckbox(name: string, index?: number) {
+    return this.getEntityCheckbox(this.entitySelector, name, index);
+  }
+
+  public async getConversationCheckboxState(name: string, index?: number) {
+    return this.getEntityCheckboxState(this.entitySelector, name, index);
+  }
 }
